Failing test for specifying node-version with pinned project.
Leverages the ability to checkout specific branches in
`action/checkout@v2` to checkout the main repo (`volta-cli/action`) into
a subfolder (`./action`) and an orphaned branch with only a
`package.json` that has pinned node/yarn versions into the working
directory root. This allows us to test the action as if it were _inside_
a normal project setting (aka the actions own setup does not break the
CI test).