Add `package-json-path` input to specify location of `package.json`
For projects without a `package.json` in the working directory root we
currently do not support the ability to pin specific `node-version` |
`yarn-version` | `npm-version` (we are only able to ensure that the
given versions are preinstalled, but when `node`|`yarn`|`npm` are
ultimately ran they might not use the specified). This is because we are
only able to pin if we can find the `package.json` which contains the
`volta` property (which we need to override).

This adds a new input `package-json-path` that can be used to specify
_where_ in a project the `package.json` is actually located. When not
specified, we will use `package.json` (e.g. a `package.json` in the root
of the working directory).